Internalize serializerMap in schema module
