+++ packages/hub/services/search-index.ts
@@ -1,6 +1,5 @@
 import { Compiler, makeGloballyAddressable } from '@cardstack/core/src/compiler';
 import {
-  CardComponentMetaModule,
   CardModel,
   CardSchemaModule,
   CompiledCard,
@@ -9,6 +8,7 @@ import {
   RawCard,
   Unsaved,
 } from '@cardstack/core/src/interfaces';
+import CardModelImpl from '@cardstack/core/src/card-model';
 import { RawCardDeserializer, RawCardSerializer } from '@cardstack/core/src/serializers';
 import { cardURL } from '@cardstack/core/src/utils';
 import { JS_TYPE } from '@cardstack/core/src/utils/content';
@@ -29,7 +29,6 @@ import { service } from '@cardstack/hub/services';
 
 import { transformToCommonJS } from '../utils/transforms';
 import flatMap from 'lodash/flatMap';
-import CardModelForHub from '../lib/card-model-for-hub';
 import { INSECURE_CONTEXT } from './card-service';
 
 const log = logger('hub/search-index');
@@ -273,12 +272,10 @@ class IndexerRun implements IndexerHandle {
     );
     let format: Format = 'isolated';
 
-    let componentMetaModule = definedCard.componentInfos[format].metaModule.global;
-    let componentMeta: CardComponentMetaModule = await this.fileCache.loadModule(componentMetaModule);
     let schemaModule: CardSchemaModule = await this.fileCache.loadModule(definedCard.schemaModule.global);
     let cardService = await this.cardService.as(INSECURE_CONTEXT);
 
-    let cardModel = new CardModelForHub(
+    let cardModel = new CardModelImpl(
       cardService,
       {
         type: 'loaded',
@@ -292,7 +289,6 @@ class IndexerRun implements IndexerHandle {
         schemaModuleRef: definedCard.schemaModule.global,
         schemaModule,
         componentModuleRef: definedCard.componentInfos[format].componentModule.global,
-        componentMeta,
         saveModel: cardService.saveModel.bind(cardService),
       }
     );
